Extracting "editorFactory" module
