+++ testing/tests/DevExpress.ui.widgets/filterBuilder.tests.js
@@ -60,3 +60,4 @@ QUnit.testStart(function() {
 
 require("./filterBuilderParts/commonTests.js");
 require("./filterBuilderParts/utilsTests.js");
+require("./filterBuilderParts/eventsTests.js");
